Move product config to shared
